% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export.R
\name{export_abundance}
\alias{export_abundance}
\title{Export abundance matrix from TSE}
\usage{
export_abundance(tse, file, assay.type = "counts", format = c("csv", "xlsx"))
}
\arguments{
\item{tse}{TreeSummarizedExperiment object}

\item{file}{Output file path}

\item{assay.type}{Which assay to export (default: "counts")}

\item{format}{Output format: "csv" or "xlsx"}
}
\value{
Invisible NULL
}
\description{
Export abundance matrix from TSE
}
