% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export.R
\name{export_vpftse}
\alias{export_vpftse}
\title{Export a TSE object to RDS}
\usage{
export_vpftse(tse, file)
}
\arguments{
\item{tse}{TreeSummarizedExperiment object}

\item{file}{Output file path}
}
\value{
Invisible NULL
}
\description{
Export a TSE object to RDS
}
