% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_dramv}
\alias{read_dramv}
\title{Read DRAM-v annotations}
\usage{
read_dramv(fin)
}
\arguments{
\item{fin}{DRAM-v annotations.tsv file}
}
\value{
data.frame with gene-level annotations
}
\description{
Read DRAM-v annotations
}
