% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_genomad}
\alias{read_genomad}
\title{Read geNomad virus summary}
\usage{
read_genomad(fin)
}
\arguments{
\item{fin}{geNomad virus summary TSV file}
}
\value{
data.frame
}
\description{
Read geNomad virus summary
}
