% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_phist}
\alias{read_phist}
\title{Read PHIST host predictions}
\usage{
read_phist(fin)
}
\arguments{
\item{fin}{PHIST predictions TSV file}
}
\value{
data.frame
}
\description{
Read PHIST host predictions
}
