% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_tblastx}
\alias{read_tblastx}
\title{Read a tblastx table generated by Easyfig}
\usage{
read_tblastx(fin, seq1 = 0, seq2 = 0, max_evalue = 0.001, min_bitscore = 50)
}
\arguments{
\item{fin}{`12.easyfig.out`}

\item{seq1}{Query genome name}

\item{seq2}{Target genome name}

\item{max_evalue}{E-value threshold}

\item{min_bitscore}{Bitscore threshold}
}
\value{
data.frame
}
\description{
This is for adding links to gggenomes.
}
