% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tse_virome}
\alias{tse_virome}
\title{TSE virome}
\format{
TSE
}
\source{
{Created in-house PAB profile}
data(tse_virome)
}
\usage{
tse_virome
}
\description{
Virome contig annotation in TSE format
Report ...
}
\keyword{datasets}
