% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{abundance_adjust_by_covfrac}
\alias{abundance_adjust_by_covfrac}
\title{Adjust abundance table by coverage fraction table}
\usage{
abundance_adjust_by_covfrac(df_abundance, df_covfrac, covfrac_threshold = 0.5)
}
\arguments{
\item{df_abundance}{assay(tse, "counts") table}

\item{df_covfrac}{assay(tse, "covfrac") table}

\item{covfrac_threshold}{minimum covfrac (default: 0.5)}
}
\value{
matrix
}
\description{
Adjust abundance table by coverage fraction table
}
