% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{create_vpftse}
\alias{create_vpftse}
\title{Create TSE object}
\usage{
create_vpftse(
  fin_abcount,
  fin_abtpm,
  fin_abtmm,
  fin_abcov,
  fin_taxa,
  fin_checkv,
  fin_virsorter2,
  fin_vibrant,
  fin_dvf,
  fin_replicyc,
  df_metadata = NULL
)
}
\arguments{
\item{fin_abcount}{File `abundance_contigs_count.tsv.gz` created by coverm}

\item{fin_abtpm}{File `abundance_contigs_tpm.tsv.gz` created by coverm}

\item{fin_abtmm}{File `abundance_contigs_tmm.tsv.gz` created by coverm}

\item{fin_abcov}{File `abundance_contigs_covfrac.tsv.gz` created by coverm}

\item{fin_taxa}{File `taxa_mmseqs_formatted_all.tsv` created by coverm}

\item{fin_checkv}{File `quality_summary.tsv` created by checkv}

\item{fin_virsorter2}{File `final-viral-score.tsv` created by VirSorter2}

\item{fin_vibrant}{File `VIBRANT_genome_quality_contigs.tsv` created by VIBRANT}

\item{fin_dvf}{File `dvf_virus.tsv.tsv` created by DVF}

\item{fin_replicyc}{File `putative_vcontigs_pref1.fasta.bacphlip` created by Replicyc}

\item{df_metadata}{Metadata table as data.frame}
}
\value{
TreeSummarizedExperiment object
}
\description{
Create TSE object
}
