% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{create_vpftse_vir}
\alias{create_vpftse_vir}
\title{Create TSE virome object}
\usage{
create_vpftse_vir(tse)
}
\arguments{
\item{tse}{TSE object}
}
\value{
TreeSummarizedExperiment object
}
\description{
Create TSE virome object
}
