% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export.R
\name{export_annotations}
\alias{export_annotations}
\title{Export contig annotations from TSE}
\usage{
export_annotations(tse, file, format = c("tsv", "csv"))
}
\arguments{
\item{tse}{TreeSummarizedExperiment object}

\item{file}{Output file path}

\item{format}{Output format: "tsv" or "csv"}
}
\value{
Invisible NULL
}
\description{
Export contig annotations from TSE
}
