% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_catbat}
\alias{read_catbat}
\title{Read CATBAT results}
\usage{
read_catbat(fin)
}
\arguments{
\item{fin}{CATBAT taxonomy annotation file}
}
\value{
data.frame
}
\description{
Read CATBAT results
}
