% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_checkv}
\alias{read_checkv}
\title{Read CheckV results}
\usage{
read_checkv(fin)
}
\arguments{
\item{fin}{CheckV results}
}
\value{
data.frame
}
\description{
Add "checkv_" prefix to column names.
}
