% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_dvf}
\alias{read_dvf}
\title{Read DeepVirFinder results}
\usage{
read_dvf(fin, thr_score = 0.9, thr_pvalue = 0.01, thr_qvalue = 0.01)
}
\arguments{
\item{fin}{file of DeepVirFinder results}

\item{thr_score}{threshold of score}

\item{thr_pvalue}{threshold of pvalue}

\item{thr_qvalue}{threshold of qvalue}
}
\value{
data.frame
}
\description{
Read DeepVirFinder results
}
