% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_pharokka}
\alias{read_pharokka}
\title{Read pharokka functional annotations}
\usage{
read_pharokka(fin)
}
\arguments{
\item{fin}{pharokka CDS output TSV file}
}
\value{
data.frame with gene-level annotations
}
\description{
Read pharokka functional annotations
}
