% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_replicyc}
\alias{read_replicyc}
\title{Read replication cycle prediction results from Bacphlip}
\usage{
read_replicyc(fin, tool = "bacphlip", version = "v1")
}
\arguments{
\item{fin}{file path of Bacphlip}

\item{tool}{tool name}

\item{version}{tool version}
}
\value{
data.frame
}
\description{
Read replication cycle prediction results from Bacphlip
}
