% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_taxonomy}
\alias{read_taxonomy}
\title{Read taxonomy table}
\usage{
read_taxonomy(fin, tool = "mmseqs", version = "v1")
}
\arguments{
\item{fin}{taxonomy annotation file}

\item{tool}{tool used to annotate taxonomy}

\item{version}{version of the tool used to annotate taxonomy}
}
\value{
data.frame
}
\description{
Read taxonomy table
}
