% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_taxonomy2}
\alias{read_taxonomy2}
\title{Read mmseqs2 taxonomy table}
\usage{
read_taxonomy2(fin)
}
\arguments{
\item{fin}{taxonomy annotation file}
}
\value{
data.frame
}
\description{
Read mmseqs2 taxonomy table
}
