% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_virsorter2}
\alias{read_virsorter2}
\title{Read VirSorter2 results}
\usage{
read_virsorter2(fin)
}
\arguments{
\item{fin}{VirSorter2 results}
}
\value{
data.frame
}
\description{
Add "virsorter2_" prefix to column names.
}
