% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{read_vrhyme}
\alias{read_vrhyme}
\title{Read vRhyme viral bin assignments}
\usage{
read_vrhyme(fin)
}
\arguments{
\item{fin}{vRhyme bin-to-contig mapping file}
}
\value{
data.frame
}
\description{
Read vRhyme viral bin assignments
}
