% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readvpf.R
\name{rpb2bpb}
\alias{rpb2bpb}
\title{Convert reads_per_base to base_per_base (depth)}
\usage{
rpb2bpb(fin, reads_len = 150)
}
\arguments{
\item{fin}{path of reads_per_base file}

\item{reads_len}{length of clean reads}
}
\value{
matrix
}
\description{
Convert reads_per_base to base_per_base (depth)
}
