% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfkit-package.R
\docType{package}
\name{vpfkit-package}
\alias{vpfkit}
\alias{vpfkit-package}
\title{vpfkit: Analyzing ViroProfiler results}
\description{
An R package for analyzing ViroProfiler results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/deng-lab/vpfkit}
  \item Report bugs at \url{https://github.com/deng-lab/vpfkit/issues}
}

}
\author{
\strong{Maintainer}: Jinlong Ru \email{jinlong.ru@gmail.com} (\href{https://orcid.org/0000-0002-6757-6018}{ORCID})

Authors:
\itemize{
  \item Jinlong Ru \email{jinlong.ru@gmail.com} (\href{https://orcid.org/0000-0002-6757-6018}{ORCID})
}

}
\keyword{internal}
